Update class.Images.php
Add "Last Modified" and (optionally, from calling script) "Expires" headers for better browser caching. Default "Expires" is 30 days. If "Last Modified" of source file is not ascertainable, then the "Last Modified" header is not set. This will disable browser caching (which is no change from the result of previous versions of this code).